refactor: LadderGame 진행을 Ladder에서 수행
