+++ src/main/java/laddergame/LadderGameApplication.java
@@ -10,14 +10,11 @@ public class LadderGameApplication {
         try {
             Players players = new Players(InputView.inputPlayerNames());
             Rewards rewards = new Rewards(InputView.inputRewardNames());
-            LadderGame ladderGame = new LadderGame(players.size(),
-                    InputView.inputHeight(),
-                    new RandomCreateRule());
+            Ladder ladder = new Ladder(players.size(), InputView.inputHeight(), new RandomCreateRule());
 
-            OutputView.outputLadderGame(ladderGame, players, rewards);
+            LadderGameResult ladderGameResult = ladder.startGame(players, rewards);
 
-            LadderGameResult ladderGameResult = ladderGame.startGame(players, rewards);
-
+            OutputView.outputLadderGame(ladder, players, rewards);
             OutputView.outputLadderGameResult(InputView.inputWantResult(), ladderGameResult);
         } catch (Exception e) {
             System.err.println(e.getMessage());
