+++ src/test/java/laddergame/domain/LadderTest.java
@@ -28,4 +28,16 @@ class LadderTest {
         assertThat(ladder.takeLadder(0)).isEqualTo(1);
         assertThat(ladder.takeLadder(1)).isEqualTo(0);
     }
+
+    @Test
+    void 게임이_올바르게_진행되는지_테스트() {
+        Players players = new Players(Arrays.asList("pobi", "cu"));
+        Rewards rewards = new Rewards(Arrays.asList("1000", "100"));
+        Ladder ladder = new Ladder(2, 5, new AlwaysCreateRule());
+
+        LadderGameResult ladderGameResult = ladder.startGame(players, rewards);
+
+        assertThat(ladderGameResult.result("pobi").getName()).isEqualTo("100");
+        assertThat(ladderGameResult.result("cu").getName()).isEqualTo("1000");
+    }
 }
